+++ CHANGELOG.md
@@ -30,6 +30,7 @@ controller or added a new module you need to rename `feature` to `component`.
 - **decidim-meetings**: [\#3022](https://github.com/decidim/decidim/pull/3022)
   + Introduce `ViewModel` and `Cells`. Add `MeetingCell` with two variations: `MeetingMCell` and `MeetingListItemCell`.
   + Add the `card` attribute to the component's manifest `shared/author_reference` partials.
+- **decidim-surveys**: Add rich text description to questions [\#3066](https://github.com/decidim/decidim/pull/3066).
 
 **Changed**: